+++ src/index.js
@@ -2,17 +2,43 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
-import App from "./App";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-// import Signup from "./pages/Signup";
-// import Signin from "./pages/Signin";
-// import Todo from "./pages/Todo";
-// import Root from "./pages/Root";
+import App from "./App";
+import Home from "./pages/Home";
+import Signup from "./pages/Signup";
+import Signin from "./pages/Signin";
+import Todo from "./pages/Todo";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <App />,
+    children: [
+      {
+        path: "/",
+        element: <Home />
+      },
+      {
+        path: "signup",
+        element: <Signup />
+      },
+      {
+        path: "signin",
+        element: <Signin />
+      },
+      {
+        path: "todo",
+        element: <Todo />
+      }
+    ]
+  }
+]);
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <App />
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
